axi_dmac: Silence Quartus warnings about implicit truncation
Quartus warns for expressions like `x <= x + 1; that there is a implicit
truncation because 1 is a 32 bit constant and x is less than 32 bit.

It really shouldn't warn about this since it is easy enough to see that 1
doesn't need the full 32 bits. But its also easy to fix the warning by
replacing 1 with 1'b1.

Lets do that to have less false positive warnings in the log.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>