m2k: Temporarily removed axi_rd_wr_combiner modules
The axi_rd_wr_combiner modules main purpose is to reduce the overall
utilization. By connecting a DMA that reads from memory and another one
that writes into memory to the same slave interface of the HP interconnect,
avoiding in this way the use of another slave interface on the same
interconnect.
The axi_dmac got a lot of new features, lately. Because of this, the
axi_rd_wr_combiner must be updated. Until then, as a workaround we
will remove it from the system, increasing the system utilization by 2%.