Revert "axi_ad9162: Integrate generic JESD204 TPL"
This reverts commit 6afd1fd9c83e5032c75924d92c7bcd88c86427bb.

The generic DAC TPL and Cordic DDS module do not support complex mode, which is
required by the AD9162.

Because implementing complex mode support would effect the generic source base,
and we are just about closing the hdl_2018_r2 release, just revert the generic
DAC TPL patch, and use the old transport layer.

The patch which will add complex samples support for both the DDS and TPL DAC
will be added to the next release.