ad_dds: Implement complex mode
Several DACs have a so called complex mode, when a device expect complex samples
(sequence of I/Q pairs) as its digital input, and the mixing of these
samples are done inside the device, before the actual digital to analog
conversion.

This patch modifies the DDS to be able to generate complex samples.