adrv9009_zu11eg_som: Reset data paths before capture
Reset the upack/cpack modules before we start a TX/RX capture.

The reset is launched at the positive edge of SYS_REF, and is active for
one core clock period.
The RX capture and TX DDS is launched after exactly three core clock cycle of
the positive edge of SYSREF.