ad77681evb: Set spi_clk to 40MHz
The SPI Engine clock should be two times the maximum SCLK supported by
the device. This way we can have the biggest timing margins inside the
SPI Engine core and, at the same time, the software cand set any SPI
clock which is less or equal than the maximum supported clock.

As the maximum supported clock of the AD7768-1 device is 20MHz, the spi
clock should be set to 40MHz. If the value of the pre-scaler register is
set to 0 from the software, this 40MHz spi clock will be divided by 2,
resulting in the desired value of 20MHz.