partial_reconfig: Modify upack and cpack connections
Modified the upack and cpack connections. Placed the partial reconfigurable
module between the upack and the fifo on the Tx path and between the pack
and the fifo on the Rx path.
Modified the adi_project_impl function in the adi_project_xilinx.tcl file
so that it creates .xsa files instead of .hdf.

Signed-off-by: Dan Hotoleanu <dan.hotoleanu@analog.com>