axi_dac_interpolate: Fix low samplerate pulse
The low samplerate pulse, present in the beginning of the transmission could
be replicated only with libm2k after a DAC calibration procedure.

The cause is that after the DMA channel is stopped(from cyclic transfer)
the last data will remain on the data bus. The DAC acts as the master
entity in the data paths, after its initialization it will continuously
require data. Since the filters in the axi_dac_interpolate are enabled
by the DAC which always requires data and the DMA provides a random
sample from the moment it was stopped, this allows the filtering path to
be filled with a constant value. The calibration procedure uses a
constant level(DMA in cyclic mode) of a high amplitude. At low sample
rates the filtering path pushes data according to the requested data rate
this results in a pulse at the beginning of the transmission at low
sample rates.

This scenario was created by three changes:
-DMA interface changed from fifo to streaming
-fixing the synchronized start of the two DAC channels
-adding last sample hold support