adi_board.tcl: add a prefix to segment names in ad_cpu_interconnect
This adds an optional prefix to be applied to the segment to avoid names clashes. This is used when creating a block design in the root, and then moved to a hierarchical cell using move_bd_cells. Without the prefix, a second instance of this block design will have segment name conflicts because it uses the instance name before moving.