axi_adrv9001: Change the DDS sync structure
The DDS for each channel was synchronized by the main channel.
One problem with this aporoach is that when a user sets a DDS that
is not from the main channel the sinchronization does not happend.
This behavior is not user friendly in IIO-Oscilloscope or within other
configuration methods.

This commit keeps all channels in sync by triggering the sync on all
channels from each individual channel.