projects/scripts: Raising error at version check.
Added returning with error because Propel Builder (propelbld) does
not exit with the exit command if it is run with a tcl script as
argument.
With this modification propelbld does exit, but doesn't return
with error so the make script will write OK for the run,
instead of FAILED, but at least the Radiant project can not be
created, becouse the project folder does not exist.

Signed-off-by: Villyam <Vilmoscsaba.Jozsa@analog.com>