scripts: xilinx: always call library.mk on deps
Library dependencies are managed in library.mk, this commit ensures
that library.mk manage and skip them by setting a dummy FORCE target.
When building a project, it is desired to rebuild all libraries that
have been modified.
Previously, the solution placed component.xml files as dependencies,
skipping these targets if the file already exists, but ignored changes
to them.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>