projects/scripts: Updating adi_project_lattice_bd.tcl script.
Workaround for keeping the configured IP folders in Propel Builder 2023.2
command line version.
The 'sbp_design save' doesn't saves the temporary .lib folder to lib folder,
instead deletes if there is anything in lib folder.
I am copying the content of .lib (the configured IP cores) to lib after save
to keep the configured IP cores for the Radiant project.
Also generating the bsp after copying the configured IP cores to lib folder
because that's also based on the content of lib folder.

Signed-off-by: Villyam <Vilmoscsaba.Jozsa@analog.com>