spi_engine: scripts: Fix offload clock domain
At the configuration instantiated by the script, ctrl_clk and spi_clk are
synchronous, so the ASYNC_SPI_CLK should be Low, regardless of the AXI
bus speed.
The output of the ctrl_offload intf in the axi_spi_engine core is
already fully in the spi_clk domain, when ASYNC_SPI_CLK=1 at that core.
This change improves the offload performance.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>