hdl: AD738x Four SDI Lines
With the help of the ALERT_SPI_N parameter, we can do a build
in which we choose what functionality want for SDOD/ALERT pin.
The number of sdi lines is set to 4. When Alert function is enabled
data are present on sdia line. This is specific to the
ad7380-4 chip

Signed-off-by: Pop Ioan Daniel <pop.ioan-daniel@analog.com>