library: spi_execution: Debug counter
By performing a test in the testbench, a bug was
discovered regarding a delay set through the sleep_time register.
This delay differs depending on how many bits we set to be transferred,
although we do not want this. By changing the line of code above the
bug disappeared. But we don't know the implications it has.

Signed-off-by: Pop Ioan Daniel <pop.ioan-daniel@analog.com>