fmcomms2: Fix register not packing into IOB
Starting on Vivado 2023.2, multiple nets in between a register and an
OBUF causes the packing into IOB to fail, even if the others nets are
disconnected.
Remove spi_clk_i/o loopback on the system_top to ensure only one net in
this path, since the target register is instantiated at IP level.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>