Add backstage catalog generator
Creates components for HDL libraries and projects,
leverages the Doctools HDL parser to create the dependency hierarchy.
Descriptions and tags are obtained from the documentation.
Part names in the :adi: role are converted into tags.
The CI only commits when there are changes, and always amends the
commit.

After PR approval, update the CI to target refs/heads/main.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>