pulsar_adc: add split transfer functionality
add a parameter, SPLIT_TRANSFER_SIZE, to enable and control split transfers. If
SPLIT_TRANSFER_SIZE is set to 0 (default), the project behaves as usual (no
split transfers). Otherwise, an asymetric fifo is used to stitch together 2
transfers before sending them to the DMA, SPLIT_TRANSFER_SIZE being used to
indicate the size of each transfer in the split.

Signed-off-by: Laez Barbosa <laez.barbosa@analog.com>