SPI Engine: simplify interconnect
Make the interconnect use an external signal for switching between
sources, so it becomes just a collection of muxes, reducing latency.
This signal is generated based on the software-controlled offload
enable, but is sourced from the offload module to avoid unnecessary
CDCs.

This commit makes the use of SYNC optional.

Signed-off-by: Laez Barbosa <laez.barbosa@analog.com>