i3c_controller: Fix Start bit after direct Stop
There is no guarantee that the device driver will use the same xfer
(to yield repeated start (Sr)), for multiple concurrent transfer.
Instead, it may queue multiple xfers which generate
(...Stop)+(Start...).
This case would cause the bit modulation to yield a
[Stop+RepeatedStart], generating an "extra" clock bit.
Check for the bit-mod if the current state is Stop, if so, clear
the autoset Sr state reg, ensuring a [Stop+Start].

Signed-off-by: Jorge Marques <jorge.marques@analog.com>