FIX: single and quad SPI fixes
Dual SPI is still an open issue. Even changing the 0f reg
from 0xh0 to 0xh1 it is still working as single mode.
This was tested with EVAL-SDP-CHA1Z.

sclk frequency lowered from 66 MHz to 16.5 MHz due to the evalboard
limitation for the config mode. Streaming should work with 66 MHz.

Signed-off-by: <carlos.souza@analog.com>