docs: Drop wavedrom, reorder requirements
Sphinx-wavedrom, cairocffi, CairoSVG don't seem to be tested
continuously, which leads to fatal bugs reaching production.
The extension in particular has two design oversights:

* It renders on the client side, requiring wavedrom.js to be
  fetched from a different from a different domain, slowing
  down or even preventing the documentation if the domain is
  offline (cause of a previous problem).
* Despite this, it still imports cairo during build
  (the root cause of the current issue)

Drop it in favor of just committing the wavedrom source (commented out
in the page or as a text file in the same path) and the rendered svg.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>