scripts: adi_ip_xilinx: Fix fall through, missing close quote
Automatic parameters from adi_xilinx_device_info_enc are injected in
adi_ip_xilinx to codegen the bd/bd.tcl files.
First, it adds the items that are lib user parameters to an array,
suffixed with backslash, then replaces the last item backslash from
info_enc with the closing double quote.
If FPGA_TECHNOLOGY is not in the IP, no double quote is inserted,
and the string is never closed on bd.tcl.
Assign the matched user parameter to an auxiliary variable and
use that on the replace instead.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>