SPI Engine: simplify interconnect (#1502)
Make the interconnect use an external signal for switching between
sources, so it becomes just a collection of muxes, reducing latency.
This signal is generated based on the software-controlled offload
enable, but is sourced from the offload module to avoid unnecessary
CDCs. 

Also updates axi_ad5766 to introduce the interfaces used for 
controlling the interconnect.

This commit makes the use of SYNC optional, unless the active signal
from spi_engine_execution is needed.

Signed-off-by: Laez Barbosa <laez.barbosa@analog.com>

* axi_ad5766: add interconnect control

Signed-off-by: Laez Barbosa <laez.barbosa@analog.com>

---------

Signed-off-by: Laez Barbosa <laez.barbosa@analog.com>