library/spi_engine: fix PCORE_VERSION (#1637)
Commit 24f2454ea ("SPI Engine: simplify interconnect (#1502)") bumped the patch version in
the documentation but missed changing it in PCORE_VERSION which is the value used by the
version register. It also made it so that the SYNC command is no longer required. This is actually a "feature" we want to test for. So that commit should have bumped the minor version instead of the patch version. Bump the version
now so that we can test for this feature in the software driver.

Signed-off-by: David Lechner <dlechner@baylibre.com>