axi_dmac: End cyclic transfers by lowering CYCLIC
This commit updates the DMA to end the cyclic transfers by lowering
the CYCLIC flag when such a transfer is in progress. This allows
for a smooth transition between different transfers, without having
to disable the DMA.
The DMA will also generate 1 SOT interrupt and 1 EOT interrupt to
mark the limits of the cyclic transfer.

Signed-off-by: Ionut Podgoreanu <ionut.podgoreanu@analog.com>