library/spi_engine: Update for SDO extension
* Extend SDO support to 8 (symmetrical with SDI support);
* Update SDI to use asymmetrical FIFO;
* Insert symmetrical FIFO for the SDO;
* Insert SDI lane mask configuration instruction to reg 3'b011;
* Insert SDO lane mask configuration instruction to reg 3'b100;
* Insert offload active interface for interconnect and execution;
* Remove register 8'h3b from spi engine regmap;
* Improvements on the critical paths of the execution module.

Prefetching on offload work iff all lanes are active.

Signed-off-by: Carlos Souza <carlos.souza@analog.com>

library/spi_engine: update spi lane mask instruction

There is a different register for SDI lane mask and SDO
lane mask;
Update the NUM_OF_SDI parameter to NUM_OF_SDIO parameter to reflect
that both SDI and SDO are symmetrical. The control over them is
through the SDI/SDO lane masks.

Inserts a ready signal for when the valid_indices
has finished its inner logic. This avoid the possible
issue where the latency of the command is smaller than
this logic. This could only happen in the FIFO mode.

Signed-off-by: Carlos Souza <carlos.souza@analog.com>

library/spi_engine: fix execution module

It was missing the sdi_lane_mask register.
That was not affecting execution because the only
logic requiring this register is on the axi_spi_engine.v

Signed-off-by: Carlos Souza <carlos.souza@analog.com>

Update comments for index_ready

Signed-off-by: Laez Barbosa <laez.barbosa@analog.com>

fix valid_index

Previous code would create issues for lane masks different from
ALL_ACTIVE_LANE_MASK, due to a pipelining bug. Although the currently officially
supported behavior is only either one lane active or all lanes active, this is a
very simple change that helps us prepare in case we want to support other lanes
in the future, and removes some extra logic (a comparison and a mux), which can
improve timing.

Signed-off-by: Laez Barbosa <laez.barbosa@analog.com>

fix backpressure condition on sdi mask change instruction

Signed-off-by: Laez Barbosa <laez.barbosa@analog.com>

(WIP) temporary workaround for sdi_fifo_level

Introduces logic to keep track of the number of elements in
the SDI FIFO. This is due to the asymmetric FIFO implementation
not including the logic for this, and instead outputting the
level of one of its internal FIFOs.

Signed-off-by: Laez Barbosa <laez.barbosa@analog.com>

update initial value of sdi lane mask to match early fifo mode behavior

Signed-off-by: Laez Barbosa <laez.barbosa@analog.com>

library/spi_engine: fix execution

Fixes the bug related to the sdo lane config instruction, which
was high all the time and not for a single cycle. That was causing
the data_assemble module to be rebuilding the lookup table, and
consequently affecting the output of the offload mode.

Removed lane_index_d because it is not necessary anymore. Both FIFO
and OFFLOAD modes are using the valid_indices lookup table.

Signed-off-by: Carlos Souza <carlos.souza@analog.com>

library/spi_engine: update version

The modifications performed in this library are not backwards compatible. So
this version is Spi Engine 2.0.

Signed-off-by: Carlos Souza <carlos.souza@analog.com>

library/spi_engine: Update for SDO extension

* Extend SDO support to 8 (symmetrical with SDI support);
* Update SDI to use asymmetrical FIFO;
* Insert symmetrical FIFO for the SDO;
* Insert SDI lane mask configuration instruction to reg 3'b011;
* Insert SDO lane mask configuration instruction to reg 3'b100;
* Insert offload active interface for interconnect and execution;
* Remove register 8'h3b from spi engine regmap;
* Improvements on the critical paths of the execution module.

Prefetching on offload work iff all lanes are active.

Signed-off-by: Carlos Souza <carlos.souza@analog.com>

library/spi_engine: merge fix

merge fix to the axi_spi_engine.v

Signed-off-by: Carlos Souza <carlos.souza@analog.com>